move and rename function for setting model for invite
