+++ frontend/src/components/ActiveUsers/ActiveUserItem.tsx
@@ -1,6 +1,9 @@
+/**
+ * Active user list item
+ */
 import React from "react";
 import { User } from "../../app/features/types";
-import { handleInviteToChatClick } from '../../services/socket/publishers';
+import { setModalInviteToChat } from '../../util/middleware/socketActions/modal';
 
 interface ActiveUserItemProps {
   user: User;
@@ -11,7 +14,7 @@ const ActiveUserItem: React.FC<ActiveUserItemProps> = ({ user }) => {
   return (
     <div className="panel-block is-active is-flex is-flex-direction-row is-justify-content-center">
       <p className="mx-4">{user.username}</p>
-      <button onClick={() => handleInviteToChatClick(user.socketId, user.username)} className="button is-link is-light">Invite to Chat</button>
+      <button onClick={() => setModalInviteToChat(user.socketId, user.username)} className="button is-link is-light">Invite to Chat</button>
     </div>
   )
 }
