+++ frontend/src/services/socket/publishers.ts
@@ -1,21 +1,6 @@
 import { socket } from "../../context/socket"
 import { setNotification } from "../../app/features/notificationSlice"
-import { setModal } from "../../app/features/modalSlice";
 import { store } from "../../app/store";
-
-export const handleInviteToChatClick = (peerId: string, peerUsername: string) => {
-  const modalData = {
-    modalName: 'send chat invite',
-    modalContent: `Would you like to invite ${peerUsername} to private chat?`,
-    confirmBtnText: 'Yes, send invite.',
-    declineBtnText: 'No, cancel invite.',
-    isActive: true,
-    peerId,
-    socketEvent: 'invite private chat'
-  }
-
-  store.dispatch(setModal(modalData))
-}
 
 /**
  * Send socket event to send invitation request to peer
